chore[package]: rename coordinate
