+++ src/main/java/janggi/domain/board/setup/BoardSetUp.java
@@ -1,6 +1,6 @@
 package janggi.domain.board.setup;
 
-import janggi.domain.coodinate.Point;
+import janggi.domain.coordinate.Point;
 import janggi.domain.piece.Piece;
 import janggi.domain.side.Side;
 import java.util.Map;
