+++ src/main/java/janggi/domain/coordinate/FixedPathStrategy.java
@@ -1,4 +1,4 @@
-package janggi.domain.coodinate;
+package janggi.domain.coordinate;
 
 import java.util.ArrayList;
 import java.util.List;
