+++ src/main/java/janggi/domain/piece/Empty.java
@@ -1,8 +1,8 @@
 package janggi.domain.piece;
 
 import janggi.domain.board.Board;
-import janggi.domain.coodinate.Path;
-import janggi.domain.coodinate.Point;
+import janggi.domain.coordinate.Path;
+import janggi.domain.coordinate.Point;
 import janggi.domain.side.Side;
 import java.util.List;
